requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==5.0
django-crispy-forms==2.1
django-decouple==2.1
django-dotenv==1.4.2
django-unused-media==0.2.2
mysqlclient==2.2.1
Pillow==10.1.0
python-decouple==3.8
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
